+++ src/PieCrust/Server/PieCrustServer.php
@@ -169,19 +169,27 @@ class PieCrustServer
             $pieCrust->getConfig()->setValue('site/cache_time', false);
             $pieCrust->getConfig()->setValue('server/is_hosting', true);
 
-            // Apply the specified configuration variant, if any. Otherwise,
-            // use the default variant.
-            $isDefault = false;
+            // New way: apply the `server` variant.
+            // Old way: apply the specified variant, or the default one. Warn about deprecation.
             $variantName = $this->options['config_variant'];
-            if (!$variantName)
+            if ($variantName)
             {
-                $isDefault = true;
-                $variantName = 'default';
+                $this->logger->warning("The `--config` parameter has been moved to a global parameter (specified before the command).");
+                $this->pieCrust->getConfig()->applyVariant("server/config_variants/{$variantName}");
+                $this->logger->warning("Variant '{$variantName}' has been applied, but will need to be moved to the new `variants` section of the site configuration.");
             }
-            $pieCrust->getConfig()->applyVariant(
-                "server/config_variants/{$variantName}",
-                !$isDefault
-            );
+            else
+            {
+                if ($this->pieCrust->getConfig()->hasValue("server/config_variants/default"))
+                {
+                    $this->pieCrust->getConfig()->applyVariant("server/config_variants/default");
+                    $this->logger->warning("The default server configuration variant has been applied, but will need to be moved into the new `variants/server` section of the site configuration.");
+                }
+                else
+                {
+                    $this->pieCrust->getConfig()->applyVariant("variants/server", false);
+                }
+            }
         }
         catch (Exception $e)
         {